Revert "Version 0.0.10"
