+++ Tests/Foundation/CharacterExtensionTest.swift
@@ -38,6 +38,21 @@ class CharacterExtensionTest: XCTestCase {
         XCTAssertFalse(Character("s").isUppercased)
     }
     
+    func test_int() {
+        XCTAssertEqual(Character("1").int, 1)
+        XCTAssertNil(Character("s").int)
+    }
+    
+    func test_float() {
+        XCTAssertEqual(Character("1").float, Float(1.0))
+        XCTAssertNil(Character("s").float)
+    }
+    
+    func test_double() {
+        XCTAssertEqual(Character("6").double, Double(6.0))
+        XCTAssertNil(Character("s").double)
+    }
+    
     func test_string() {
         XCTAssertEqual(Character("S").string, String("S"))
     }
